added settings sagas file and moved settings dir to scenes dir
