+++ src/scenes/AuthenticatedView.js
@@ -1,7 +1,7 @@
 import React from "react";
 import {createStackNavigator} from "react-navigation";
 import HomeView from "./HomeView";
-import SettingsView from "./user/settings/SettingsView";
+import SettingsView from "./settings/SettingsView";
 import {colors} from "../styles";
 
 const routeConfig = {
