+++ src/scenes/settings/link/styles/index.js
@@ -1,5 +1,5 @@
 import {StyleSheet} from "react-native";
-import {colors} from "../../../../../styles";
+import {colors} from "../../../../styles";
 
 const linkComponentStyles = StyleSheet.create({
     linkComponentContainer: {
