+++ src/scenes/settings/styles.js
@@ -1,5 +1,5 @@
 import {StyleSheet} from "react-native";
-import {colors} from "../../../styles";
+import {colors} from "../../styles";
 
 const styles = StyleSheet.create({
     container: {
